+++ packages/brick_offline_first/CHANGELOG.md
@@ -6,7 +6,7 @@
 * Priority for the next job to process from the queue - when processing requests in serial - has changed from `'$HTTP_JOBS_CREATED_AT_COLUMN ASC, $HTTP_JOBS_ATTEMPTS_COLUMN DESC, $HTTP_JOBS_UPDATED_AT ASC'` to `'$HTTP_JOBS_CREATED_AT_COLUMN ASC'`; this uses the job column introduced in 0.0.7 (26 May 2020) and will not affect any implementations using 0.0.7 or higher.
 * `RequestSqliteCache` no longer queries cached requests based on headers; requests are rediscovered based on their encoding, URL, request method, and body. Rehydrated (reattempted) requests will be hydrated with headers from the original request.
 * Field types in models `Set<Future<OfflineFirstModel>>`, `List<Future<OfflineFirstModel>>`, and `Future<OfflineFirstModel>` are no longer supported. Instead, use `Set<OfflineFirstModel>`, `List<OfflineFirstModel>`, and `OfflineFirstModel` (the adapters will `await` each).
-* `StubOfflineFirstWithRest.client` is no longer accessible as mockito has changed `any` to be nullable. Instead, the `http` package's `testing` is used to stub responses from declared endpoints. A client should not be explicitly declared when using `StubOfflineFirstWithRest`.
+* `StubOfflineFirstWithRest` is no longer supported. SQLiteFFI has satisfied much of the original stubbing required for this class, and http's testing.dart library is sufficient to not require mockito. Instead, pass `stubRestClient` to your `RestProvider#client` with a map of endpoints to response values. `StubOfflineFirstWithRestModel` has also changed significantly: custom repsonses can be provided instead of using a JSON file, and the endpoints are no longer provided as a list.
 
 ## 0.1.2